Changed types of __size_t and __ssize_t to long to avoid problems resulting from size_t and ssize_t having different sizes of x86-64 (long-sized size_t was defined in GCC headers, while ssize_t in AROS headers was int-sized). It breaks binary compatibility of x86-64 AROS, but I guess there's not much people using it anyway.
git-svn-id: https://svn.aros.org/svn/aros/trunk@30732 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ compiler/clib/include/sys/_types.h
@@ -50,9 +50,9 @@ typedef __intptr_t              __pid_t;
 typedef __intptr_t              __ptrdiff_t;        /* XXX __intptr_t? */
 typedef __uint8_t               __sa_family_t;
 #undef __size_t
-typedef unsigned int            __size_t;           /* XXX Large Files? */
+typedef unsigned long           __size_t;           /* XXX Large Files? */
 typedef __uint32_t              __socklen_t;
-typedef int                     __ssize_t;
+typedef long                    __ssize_t;
 typedef __int32_t               __suseconds_t;
 typedef __int32_t               __time_t;           /* XXX Limiting? */
 typedef int                     __timer_t;
